Recover Board Star source

+++ include/REL/w01Dll_world01.h
@@ -34,7 +34,7 @@ void mbev_ShopExInit(int dataNum, void (*hook)(int modelId, int shopNo));
 void mbev_ShopBackCreate(int dataNum, int motDataNum, int motNo, BOOL linkF);
 
 void mbStarMoveHookSet(void (*hook)(void));
-void mbStarMasuNextSet(s16 masuId);
+void mbStarMasuNextSet(int masuId);
 
 void mbTelopTimeChangeCreate(void);
 BOOL mbTelopTimeChangeCheck(void);
